requirements.txt
Django
djangorestframework
requests
redis
psycopg2